Scrappy_AI/requirements.txt
streamlit
selenium
beautifulsoup4
lxml
html5lib
python-dotenv
google-genai